Arregla errores con la generacion de datos

+++ include/Programa.h
@@ -18,7 +18,7 @@ struct Programa
         int numero1;
         int indiceOperacion;
         int numero2;
-        int tiempoEstimado;
+        float tiempoEstimado;
         int ID;
         float tiempoTranscurrido;
 };
